Restructured the EMSignal class into the Amorphous2d class.  Got rid of some mask operations that were already implemented due to the use of numpy arrays.  Seperated the MaskSlicer and Mask passer into their own classes. Might have to adjust that again when the lazy loading is looked at in more detail
